+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
- * The `projected_embeddings_file.csv` of project stages has been renamed to `projected_reduced_embeddings_file.csv`. For backwards compatibility, `projected_embeddings_file.csv` remains as symlink to `projected_reduced_embeddings_file.csv` 
+ * The `projected_embeddings_file.csv` of project stages has been renamed to `projected_reduced_embeddings_file.h5`. For backwards compatibility, `projected_embeddings_file.csv` is still written.
+ * The `projected_embeddings_file` parameter of visualize stages has been renamed to `projected_reduced_embeddings_file` and takes an h5 file. For backwards compatibility, `projected_embeddings_file` and csv files are still accepted.
  * Added the pb_tucker model as project stage. Tucker is a contrastive learning model trained to distinguish CATH superfamilies. It consumes prottrans_bert_bfd embeddings and reduces the embedding dimensionality from 1024 to 128. See https://www.biorxiv.org/content/10.1101/2021.01.21.427551v1
 
 ## v0.1.7